+++ GOOUT/SceneDelegate.swift
@@ -16,7 +16,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
 
-        let mainViewController = InquiryByGradeViewController()
+        let mainViewController = SigninViewController()
 
         let navigationController = UINavigationController(rootViewController: mainViewController)
